refactor(audit): extract render and policy helpers, name anonymous struct
Split audit.go (1349→997 lines) into focused files:
- audit_render.go: all terminal output (printSkillResult*, buildAuditSummaryLines, presentAuditResults, printAuditHelp)
- audit_policy.go: policy label/colorize helpers shared by CLI and TUI

Additional cleanups:
- Replace anonymous struct{name,path} with named auditSkillRef
- Move modeString to mode.go, pathExists to helpers.go
- Fix recountSummary missing PolicyAnalyzers preservation after dedupe
- Replace 3 hardcoded ANSI escape codes with ui.Green/ui.Gray/ui.Reset
